refact: move SerializableList to list.serializable
