+++ loadable/src/main/java/com/jeanbarrossilva/loadable/list/serializable/Collection.extensions.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.loadable.list
+package com.jeanbarrossilva.loadable.list.serializable
 
 import java.io.Serializable
